requirements.txt
ultralytics==8.3.102
streamlit==1.44.1
gdown
streamlit
Pillow
opencv-python
numpy
